move constantinfo.ts to util folders and move header navbar to constantinfo file
