+++ src/pages/Dating/ConsiderPet.tsx
@@ -2,7 +2,7 @@ import React, { Dispatch, SetStateAction, useEffect, useState } from "react";
 import emailjs from "emailjs-com";
 import { useSelector, useDispatch } from "react-redux";
 import styled from "styled-components";
-import { area, shelterInfo } from "./ConstantInfo";
+import { area, shelterInfo } from "../../utils/ConstantInfo";
 import { Dating } from "../../reducers/dating";
 import { db, deleteFirebaseData } from "../../utils/firebase";
 import {
