+++ src/pages/Dating/UpcomingList.tsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from "react-redux";
 import styled from "styled-components";
 import { db, deleteFirebaseData } from "../../utils/firebase";
 import { collection, addDoc } from "firebase/firestore";
-import { shelterInfo } from "./ConstantInfo";
+import { shelterInfo } from "../../utils/ConstantInfo";
 import { Dating } from "../../reducers/dating";
 import { Profile } from "../../reducers/profile";
 import cutEgg from "./img/scissors.png";
